project3/requirements.txt
# Python dependencies
Flask==2.3.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
Pillow==9.5.0
Werkzeug==2.3.0